refactor(exasol): Extract and compile generator
